seperate some function
