+++ index.html
@@ -34,7 +34,7 @@
         <!-- Reload the game -->
         <button class="reload"></button>
      </section>
-     <script src="index.js"></script>
+     <script type="module" src="index.js"></script>
 </body>
 </html>
