+++ index.js
@@ -1,3 +1,5 @@
+import { gameOver } from "./eventHandlers.js";
+// import { winner } from "./choiceFunctions.js";
 const game = () => {
     let playerScore = 0;
     let computerScore = 0;
@@ -6,13 +8,14 @@ const game = () => {
 const playGame = () => {
     const rockBtn = document.querySelector('.rock');
     const paperBtn = document.querySelector('.paper');
-    const scissorBtn = document.querySelector('.scissors');
+    const scissorBtn = document.querySelector('.scissor');
     const playerOptions = [rockBtn,paperBtn,scissorBtn];
     const computerOptions = ['rock','paper','scissors']
     // function to start playing game
-    playerOptions.forEach(Option => {
-        Option.addEventListener('click',function(){
-            const moveLeft = document.querySelector('.movesleft');
+    
+    playerOptions.forEach(option => {
+        option.addEventListener('click',function(){
+            const movesLeft = document.querySelector('.movesleft');
             moves++;
             movesLeft.innerText = `Moves Left: ${10-moves}`;
             const choiceNumber = Math.floor(Math.random() * 3);
@@ -37,7 +40,7 @@ const playGame = () => {
             result.textContent = 'Tie'
         }
         else if (player == 'rock'){
-            if (compter == 'paper'){
+            if (computer == 'paper'){
                 result.textContent = 'Computer won';
                 computerScore++;
                 computerScoreBoard.textContent = computerScore;
@@ -70,38 +73,10 @@ const playGame = () => {
             } 
         }
     }
+    
     // function to run when the game is over
-    const gameOver = (playerOptions,moveLeft) => {
-        const chooseMove = document.querySelector('.move');
-        const result = document.querySelector('.result');
-        const reloadBtn = document.querySelector('.reload');
-        playerOptions.forEach(option => {
-            option.style.display = 'none';
-        })
-        chooseMove.innerText = 'Game Over!!'
-        movesLeft.style.display = 'none';
-
-        if(playerScore > computerScore){
-            result.style.fontSize = '2rem';
-            result.innerText = 'You Won The Game'
-            result.style.color = '#308D46';
-        }
-        else if(playerScore < computerScore){
-            result.style.fontSize = '2rem';
-            result.innerText = 'You Lost The Game';
-            result.style.color = 'red';
-        }
-        else{
-            result.style.fontSize = '2rem';
-            result.innerText = 'Tie';
-            result.style.color = 'grey'
-        }
-        reloadBtn.innerText = 'Restart';
-        reloadBtn.style.display = 'flex'
-        reloadBtn.addEventListener('click',() => {
-            window.location.reload();
-        })
-    } 
+    
+    
     // calling playgame function inside game
     playGame();
